refactor: split admin.js monolith into focused modules, remove jQuery
admin.js (1172 lines) split into 8 modules under resources/js/admin/:
  - auth.js     — getToken, authFetch, requireAuth, setLogout, todayIso
  - travel.js   — travel CRUD, media list, EXIF autofill, geocode map
  - posts.js    — blog post CRUD, template insertion
  - cv.js       — CV upload/delete/status
  - deploy.js   — deploy/fetch/rollback with SSE streaming
  - passkeys.js — passkey list, add, delete
  - stats.js    — visit stats display
  - notes.js    — private notes (localStorage)

admin.js is now a 17-line hub that imports and calls each module's init.

jQuery removed from admin/index.html — all admin modules use vanilla DOM
APIs (addEventListener, classList, innerHTML, .value). jQuery remains on
public pages (dom.js) and is unaffected.

escapeHtml now imported from utils/html.js instead of duplicated inline.

Closes #176
