+++ admin/index.html
@@ -234,7 +234,6 @@
     <script src="/resources/js/dev-env.js"></script>
     <script src="/resources/js/darkmode.js"></script>
     <script src="/resources/js/nav.js"></script>
-    <script src="/resources/js/jquery-3.5.1.min.js"></script>
     <script type="module" src="/resources/js/admin.js"></script>
     <script src="/resources/js/admin-init.js"></script>
 </body>
